+++ src/MTurk/MTurkClient.php
@@ -1,5 +1,5 @@
 <?php
-namespace Aws\MechanicalTurkRequesterService;
+namespace Aws\MTurk;
 
 use Aws\AwsClient;
 
@@ -84,4 +84,4 @@ use Aws\AwsClient;
  * @method \Aws\Result updateQualificationType(array $args = [])
  * @method \GuzzleHttp\Promise\Promise updateQualificationTypeAsync(array $args = [])
  */
-class MechanicalTurkRequesterServiceClient extends AwsClient {}
+class MTurkClient extends AwsClient {}